+++ packages/flutter/test/painting/image_provider_and_image_cache_test.dart
@@ -16,7 +16,7 @@ import 'mocks_for_image_cache.dart';
 void main() {
   TestRenderingFlutterBinding.ensureInitialized();
 
-  Future<ui.Codec> basicDecoder(ui.ImmutableBuffer bytes, {int? cacheWidth, int? cacheHeight, bool? allowUpscaling}) {
+  Future<ui.Codec> basicDecoder(ui.ImmutableBuffer bytes, {double? cacheWidth, double? cacheHeight, bool? allowUpscaling}) {
     return PaintingBinding.instance.instantiateImageCodecFromBuffer(bytes, cacheWidth: cacheWidth, cacheHeight: cacheHeight, allowUpscaling: allowUpscaling ?? false);
   }
